drivers: sensor: mcp9600: Added driver for mcp9600
Split driver and sample in to seperate commits.

This is the entire driver functionality to read the
temperature of a K type thermocouple. Additional code
will be needed if a different type is used or to configure
the alarm trigger functionality.

On branch sensor_mcp9600
Your branch is up to date with 'origin/sensor_mcp9600'.

Changes to be committed:
modified:   drivers/sensor/CMakeLists.txt
modified:   drivers/sensor/Kconfig
modified:   drivers/sensor/mcp9600/CMakeLists.txt
modified:   drivers/sensor/mcp9600/Kconfig
modified:   drivers/sensor/mcp9600/mcp9600.c
modified:   drivers/sensor/mcp9600/mcp9600.h
modified:   dts/bindings/sensor/microchip,mcp9600.yaml

+++ drivers/sensor/mcp9600/mcp9600.h
@@ -47,7 +47,7 @@
 
 struct mcp9600_data {
 	const struct device *i2c_master;
-	int16_t ttemp;
+	int32_t ttemp;
 };
 
 struct mcp9600_config {
